feat: Added task saving feature
